Added new task type: validate.
This task uses cd-hit-est-2d to cluster the data with a given list of official clone ID's so we can find if dart has given the wrong clone IDs or if the SNP locations are different than expected.
- Update DartUtils.py with new command line options
- Added dartqc/DartFileValidation.py to handle the new task type
- Updated bin/dartqc to interpret the new cmd line options and run the DartFileValidator

The task is expected to be run as a stand-alone operation before pre-processing.  And is mainly implemented for the prawn project which is doing many repeated batches rather than more exploratory type genotyping.